+++ custom_components/dashboardmodern/frontend/legacy/build-info.js
@@ -12,4 +12,4 @@ import "../src/sections/beta25-real-device-fixes-section.js";
 import "../src/sections/beta25-compatibility-section.js";
 import "../src/sections/beta26-real-device-stability-section.js";
 import "../src/sections/segnalazioni-section.js";
-export const BUILD_INFO = Object.freeze({"generated":true,"integrationVersion":"1.4.27","dashboardVersion":"1.4.27","moduleVersion":14,"schemaVersion":4,"date":"2026-09-13T20:09:54+00:00","commit":"ef73d7c8c48b49384c7030defcaa05e453b9939e","assetHash":"403777498d52004f"});
+export const BUILD_INFO = Object.freeze({"generated":true,"integrationVersion":"1.4.27","dashboardVersion":"1.4.27","moduleVersion":14,"schemaVersion":4,"date":"2026-09-13T21:00:14+00:00","commit":"ea0361f89f23205dd0cf1106bedf1061a67a0d17","assetHash":"d65a19205f789b24"});